+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 - On the Move Hub, `bytes()` and `bytes.find()` now truncate out-of-range
   values instead of raising `ValueError`.
 
+### Fixed
+- Fixed slow broadcasting while observing at the same time ([support#2822]).
+
+[support#2822]: https://github.com/orgs/pybricks/discussions/2822
+
 [Unreleased]: https://github.com/pybricks/pybricks-micropython/compare/v4.1.0b3...HEAD
 
 ### Fixed